requirements.txt
#apparmor==2.10.95
#apparmor.click==0.3.13.1
#apturl==0.5.2
#astroid==2.1.0
#backports.csv==1.0.7
#beautifulsoup4==4.4.1
#blinker==1.3
#Brlapi==0.6.4
#chardet==2.3.0
#checkbox-support==0.22
#click==0.4.43+16.4.20170613.0ubuntu1
#command-not-found==0.3
#cryptography==1.2.3
#defer==1.0.6
#defusedxml==0.6.0
#devscripts===2.16.2ubuntu3
#diff-match-patch==20181111
Django==2.0.9
django-appconf==1.0.3
django-authtools==1.6.0
django-autoslug-iplweb==1.9.4.dev0
django-braces==1.13.0
django-cities-light==3.5.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-environ==0.4.5
django-import-export==1.2.0
django-select2==7.1.0
easy-thumbnails==2.5
#et-xmlfile==1.0.1
#feedparser==5.1.3
#guacamole==0.9.2
#html5lib==0.999
#httplib2==0.9.1
#idna==2.0
#indicator-sound-switcher===2.1.1ubuntu0
#isort==4.3.4
#jdcal==1.4.1
#Jinja2==2.8
#language-selector==0.1
#lazy-object-proxy==1.3.1
#LibAppArmor==2.10.95
#louis==2.6.4
#lxml==3.5.0
#Magic-file-extensions==0.2
#Mako==1.0.3
#MarkupSafe==0.23
#mccabe==0.6.1
mysqlclient==1.3.7
oauthlib==1.0.3
#odfpy==1.4.0
#onboard==1.2.0
#openpyxl==2.6.2
#padme==1.1.1
#pexpect==4.0.1
#Pillow==5.3.0
#plainbox==0.25
#progressbar2==3.42.0
#ptyprocess==0.5
#pyasn1==0.1.9
#pycups==1.9.73
#pycurl==7.43.0
#pygobject==3.20.0
#PyJWT==1.7.1
#pylint==2.2.2
#pylint-django==2.0.11
#pylint-plugin-utils==0.5
#pyparsing==2.0.3
#pysmbc==1.0.15.5
#python-apt==1.1.0b1+ubuntu0.16.4.5
#python-debian==0.1.27
#python-gnupg==0.3.8
python-slugify==3.0.2
#python-systemd==231
#python-utils==2.3.0
#python3-openid==3.1.0
#pytz==2018.7
#pyxdg==0.25
#PyYAML==5.1.1
#reportlab==3.3.0
#requests==2.9.1
#requests-oauthlib==1.2.0
#screen-resolution-extra==0.0.0
#sessioninstaller==0.0.0
#six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
#sqlparse==0.2.4
#ssh-import-id==5.5
#system-image==3.1
#system-service==0.3
#tablib==0.13.0
#text-unidecode==1.2
#typed-ast==1.3.0
#ubuntu-drivers-common==0.0.0
#ufw==0.35
#unattended-upgrades==0.1
#Unidecode==1.1.1
#unity-scope-calculator==0.1
#unity-scope-chromiumbookmarks==0.1
#unity-scope-colourlovers==0.1
#unity-scope-devhelp==0.1
#unity-scope-firefoxbookmarks==0.1
#unity-scope-gdrive==0.7
#unity-scope-manpages==0.1
#unity-scope-openclipart==0.1
#unity-scope-texdoc==0.1
#unity-scope-tomboy==0.1
#unity-scope-virtualbox==0.1
#unity-scope-yelp==0.1
#unity-scope-zotero==0.1
#unity-tweak-tool==0.0.7
#urllib3==1.13.1
#usb-creator==0.3.0
#Werkzeug==0.14.1
#wrapt==1.11.1
#xdiagnose==3.8.4.1
#xkit==0.0.0
#xlrd==1.2.0
#XlsxWriter==0.7.3
#xlwt==1.3.0
